Do not remove user's role and abilities when softDeleting

+++ src/Database/Concerns/HasAbilities.php
@@ -22,7 +22,9 @@ trait HasAbilities
     public static function bootHasAbilities()
     {
         static::deleted(function ($model) {
-            $model->abilities()->detach();
+            if($model->isForceDeleting()) {
+                $model->abilities()->detach();
+            }
         });
     }
